firebase set up completed with global css styling
